fixed errors

+++ primality.py
@@ -3,7 +3,7 @@ import math
 def Primality(x):
 	if(x<=1):
 		return False
-	for i in range(2,math.sqrt(x)+1):
+	for i in range(2,int(math.sqrt(x))+1):
 		if x%i==0:
 			return False
